refactor: reorg page components into their own dir
