+++ components/pages/Bio/components/BioHeader/index.ts
