+++ components/pages/Bio/index.ts
