+++ components/pages/ResourceComponentMap.ts
