+++ components/pages/Story/layouts/default/Story.default.styles.ts
