+++ components/pages/Story/layouts/default/Story.default.tsx
