+++ components/pages/Story/layouts/default/components/StoryHeader/StoryHeader.default.styles.ts
