+++ components/pages/Story/layouts/default/styles/Story.layout.ts
