+++ components/pages/Story/layouts/default/styles/body/Story.body.story-act.ts
