+++ components/pages/Term/Term.tsx
