+++ components/pages/Term/index.ts
